types: time_point_to_string: use numeric formatting rather than chrono-format specifiers
As reported in #10867, newer versions of the fmt library
format %Y using 4-characters width, 0-padding the prefix
when needed, while older versions don't do that.

This change moves away from using %Y and friends
fmt specifiers to using explicit numeric-based formatting
conforming to ISO 8601 and making sure the year field
has at least 4 digits and is zero padded.  When
negative, the width is upped to 5 so it would show as -0001
rather than -001.

The unit test was updated respectively.

Fixes #10867

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>

Closes #10870