transport: fix enum-class which is no longer printable by fmt
Starting with version 8.1 of fmt, one can no longer print an enum
class and assume automatic conversion to the integer underlying type.
See https://github.com/fmtlib/fmt/issues/1841 for this change.

In this patch we fix the one place where we assumed this implicit
conversion works, and make the conversion explicit.

Fixes #10884

Signed-off-by: Nadav Har'El <nyh@scylladb.com>