test: perf: fix crash on exception in time_parallel_ex
The `time_parallel_ex` function creates a sharded<executor> and uses it
to run the benchmark on multiple shards in parallel. However, if the
benchmarking function throws an exception, the sharded<executor> will be
destroyed without being stopped, which triggers an assertion in
sharded<T> destructor.

This commit makes sure that the executor is stopped before being
destroyed by putting `exec.stop()` into a `seastar::defer`.