table: Make make_reader_v2_excluding_sstables() account for a possible clone of excluded files
There's an undergoing change which clones sstables from staging dir to
the base dir, allowing sstables in base dir to be compacted
indepedently. The problem is that view builder will not exclude those cloned
files when using make_reader_v2_excluding_sstables(), just the ones in
staging dir.

As a cloned file is just another link to a staging sstable, we can solve this
problem by filtering out all sstables that share the run identifier with
any of those files being explicitly excluded.

Refs #9559.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>