cql3: Keep single column restrictions map inside statement restrictions
Some parts of the code make use of a map keeping single column restrictions
for each partition key column. One of this places is inside do_filter,
so it could be a performance problem to create such a map from scratch
each time.

After adding all restrictions from the where clause the new
map is created and can be used for various purposes.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>