install-dependencies: strip libwasmtime.a from spurious debug symbols
Traditionally, libraries are installed *without* debugging symbols.
The libwasmtime.a that we install in install-dependencies is an
exception - it does include some debug symbols.

As a result, when the linker links Scylla built *without* debug symbols
(e.g., dev build mode) with this libwasmtime.a library - the result has
*some* debugging symbols, so utilities like "file" and "gdb" believe
that it has debugging symbols - although they are useless for any sort
of real debugging so any such attempt will fail in unclear ways.
This is confusing and misleading.

So in this patch, we modify install-dependencies.sh to strip this
library from its debugging symbols.

Note that for this change to take effect, the developer must re-run
install-dependencies.sh. The stripping will happen even if the library
was already downloaded in the past.

Fixes #10863

Signed-off-by: Nadav Har'El <nyh@scylladb.com>