dbuild: disable selinux instead of relabeling
By default, docker uses selinux to prevent malicious code in the container
from "escaping" and touching files outside the container - the container
is only allowed to touch files with a special selinux label, which the
outside files don't have. This is why one needs to use the ":z" option
when mounting an outside file inside docker - it asks docker to "relabel"
the directory to be usable in docker.

But this relabeling process is slow and potentially harmful if done to
large directories such as your home directory. It is also unnecessary -
we don't really need the selinux protection in dbuild. Dbuild was meant
to provide a common toolchain - not to protect the build host from a
malicious build script.

The alternative we use in this patch is "--security-opt label=disable".
This allows the container to access any file in the host filesystem,
but as usual - only if it's "mounted" into the container. All ":z"
we added in the past can be removed.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>