dbuild: use ccache by default
The dbuild container comes pre-installed with ccache, but until now
was not used by default. Making it used required an elaborate setup
described in the README.md.

In this patch, we enable it by default. The user'ss $HOME/.ccache directory
is mounted into the container, and used for the cache - which is shared
by compilations outside of dbuild.

Now that ccache is enabled by default, it is no longer an interesting
example for README.md so I switched to a different example.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>