commitlog: enforce commitlog size hard limit by default
Since the commitlog size hard limit is a failsafe mechanism,
we don't expect to ever hit it. If we do hit the limit, it means
that we have an exceptional condition in the system. Hence, the
impact of enforcing the commitlog hard limit is irrelevant.
Here we enforce the limit by default.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>