cql3: Simplify adding restrictions
The code that adds restrictions in statement_restrictions.cc
is unnecessarily convoluted.

The code to handle IS NOT NULL is actually repeated twice,
once in the constructor and once in add_is_not_restriction.
I missed this when I orignally modified this code.
There is no need to keep duplicate code, we can just
use the new add_is_not_restriction.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>