test: fix scylla-gdb scylla netw command
The command wasn't tested fully,
and when tested, started failing scylla-gdb test.

Before Raft, the list of connections to print in a single-node setup
was always empty, so a mistake in the gdb script command 'scylla netw'
didn't lead to a test failure.

With raft, there is always an RPC connection to self after initial
bootstrap, and the test begins to print connections (and fail, because
there is a bug in the printing code).

Fix that bug.